# src/main/java/com/cursorai/fitnessapp/dto/WaterEntryDto.java
package com.cursorai.fitnessapp.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class WaterEntryDto {
    private Long id;
    private LocalDateTime timestamp;
    private Integer ml;
}