# src/main/java/com/cursorai/fitnessapp/model/UserWorkout.java
package com.cursorai.fitnessapp.model;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Entity
@Table(name = "user_workouts")
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class UserWorkout {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "user_id", nullable = false)
    private User user;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "workout_id", nullable = false)
    private Workout workout;

    private LocalDateTime completedAt;
    private Integer actualDuration; // фактическая продолжительность в минутах
    private Integer actualCalories; // фактически сожженные калории
}